requirements.txt
pdfplumber
pandas
openai
python-dotenv
numpy